Let the compiler check the format strings of sys_printf and co.
sys_error, sys_printf and sys_snprintf take printf format strings, but
the compiler could not check their arguments, which let mismatches like
the %d for a U32 in the hall of fame go unnoticed. Declare them with the
printf format attribute on GCC and clang.

Fix what this finds, all in debug output: channel numbers and sizes
of type size_t printed with %d or %u, in the SDL and Rockbox sound and
memory code (the Rockbox files were checked by hand, as they can not
be built here).

+++ source/xrick/system/sysmem_rockbox.c
@@ -86,7 +86,7 @@ void sysmem_shutdown(void)
     }
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: max memory usage was %u bytes\n", maxUsedMemory);
+        sys_printf("xrick/memory: max memory usage was %lu bytes\n", (unsigned long)maxUsedMemory);
     );
 
     isMemoryInitialised = false;
@@ -116,7 +116,7 @@ void *sysmem_push(size_t size)
     stackSize += neededSize;
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: allocated %u bytes\n", neededSize);
+        sys_printf("xrick/memory: allocated %lu bytes\n", (unsigned long)neededSize);
         if (stackSize > maxUsedMemory) maxUsedMemory = stackSize;
     );
 
@@ -154,7 +154,7 @@ void sysmem_pop(void * alignedPtr)
     );
 
     IFDEBUG_MEMORY(
-        sys_printf("xrick/memory: released %u bytes\n", allocatedSize);
+        sys_printf("xrick/memory: released %lu bytes\n", (unsigned long)allocatedSize);
     );
 }
